requirements.txt
asn1crypto==0.24.0
backports.ssl-match-hostname==3.7.0.1
certifi==2019.11.28
chardet==3.0.4
pyOpenSSL==20.0.1
configparser==3.5.0b2
docker==4.1.0
entrypoints==0.3
enum34==1.1.6
idna==2.8
ipaddress==1.0.17
keyring==17.1.1
keyrings.alt==3.1.1
psutil==5.8.0
pycrypto==2.6.1
pyxdg==0.25
requests==2.22.0
SecretStorage==2.3.1
six==1.12.0
httpx==0.18.1
dnspython==2.1.0
urllib3==1.25.8
websocket-client==0.57.0
